Adds in algorithm/command selection system, along side functionality for automatic task numbering system.
Adds command line level algorithm selection, with capability of adding more algorithms or commands in future, this should be fully functional and ready to allow algorithms to be put in place for proper testing.

TaskProcess is a re-used class from the process class that should be capable of tallying an N amount of process tasks into the system. Every algorithm has been given a constructor with a TaskProcess array to facilitate further progress.